chore: Connect `move-to-postgres` to appsmithctl
